moved helpers to jest [skip ci]
