+++ src/models/event.test.ts
@@ -1,13 +1,13 @@
 import DBConnection from "../../test/util/db_connection";
-import ModelTestPass from "../../test/helpers/model/test/pass";
-import ModelTestFail from "../../test/helpers/model/test/fail";
+import ModelTestPass from "../../test/jest/helpers/model/test/pass";
+import ModelTestFail from "../../test/jest/helpers/model/test/fail";
 import BaseWorld from "../../test/jest/support/base_world";
 import Event, { EventAttributes } from "./event";
 import {
     createModels,
     loadAttributes,
-} from "../../test/helpers/model/test/setup";
-import { teardown } from "../../test/helpers/model/test/teardown";
+} from "../../test/jest/helpers/model/test/setup";
+import { teardown } from "../../test/jest/helpers/model/test/teardown";
 
 let baseWorld: BaseWorld | undefined;
