+++ src/models/permission.test.ts
@@ -1,12 +1,12 @@
 import BaseWorld from "../../test/jest/support/base_world";
 import DBConnection from "../../test/util/db_connection";
-import ModelTestPass from "../../test/helpers/model/test/pass";
+import ModelTestPass from "../../test/jest/helpers/model/test/pass";
 import Permission, { PermissionAttributes } from "./permission";
-import { teardown } from "../../test/helpers/model/test/teardown";
+import { teardown } from "../../test/jest/helpers/model/test/teardown";
 import {
     createModels,
     loadAttributes,
-} from "../../test/helpers/model/test/setup";
+} from "../../test/jest/helpers/model/test/setup";
 
 let baseWorld: BaseWorld | undefined;
