+++ src/models/quiz/attempt.test.ts
@@ -1,13 +1,13 @@
 import BaseWorld from "../../../test/jest/support/base_world";
 import DBConnection from "../../../test/util/db_connection";
-import ModelTestPass from "../../../test/helpers/model/test/pass";
-import ModelTestFail from "../../../test/helpers/model/test/fail";
+import ModelTestPass from "../../../test/jest/helpers/model/test/pass";
+import ModelTestFail from "../../../test/jest/helpers/model/test/fail";
 import QuizAttempt, { QuizAttemptAttributes } from "./attempt";
-import { teardown } from "../../../test/helpers/model/test/teardown";
+import { teardown } from "../../../test/jest/helpers/model/test/teardown";
 import {
     createModels,
     loadAttributes,
-} from "../../../test/helpers/model/test/setup";
+} from "../../../test/jest/helpers/model/test/setup";
 
 let baseWorld: BaseWorld | undefined;
