+++ src/models/quiz/section.test.ts
@@ -1,15 +1,15 @@
 import BaseWorld from "../../../test/jest/support/base_world";
 import DBConnection from "../../../test/util/db_connection";
 
-import ModelTestPass from "../../../test/helpers/model/test/pass";
-import ModelTestParentPrevent from "../../../test/helpers/model/test/parent_prevent";
+import ModelTestPass from "../../../test/jest/helpers/model/test/pass";
+import ModelTestParentPrevent from "../../../test/jest/helpers/model/test/parent_prevent";
 import Quiz, { QuizAttributes } from "./quiz";
 import QuizSection, { QuizSectionAttributes } from "./section";
-import { teardown } from "../../../test/helpers/model/test/teardown";
+import { teardown } from "../../../test/jest/helpers/model/test/teardown";
 import {
     createModels,
     loadAttributes,
-} from "../../../test/helpers/model/test/setup";
+} from "../../../test/jest/helpers/model/test/setup";
 
 let baseWorld: BaseWorld | undefined;
