+++ src/models/role.test.ts
@@ -1,13 +1,13 @@
 import BaseWorld from "../../test/jest/support/base_world";
 import DBConnection from "../../test/util/db_connection";
-import ModelTestPass from "../../test/helpers/model/test/pass";
+import ModelTestPass from "../../test/jest/helpers/model/test/pass";
 import Role, { RoleAttributes } from "./role";
 import {
     createModels,
     loadAttributes,
-} from "../../test/helpers/model/test/setup";
-import { teardown } from "../../test/helpers/model/test/teardown";
-import ModelTestParentPrevent from "../../test/helpers/model/test/parent_prevent";
+} from "../../test/jest/helpers/model/test/setup";
+import { teardown } from "../../test/jest/helpers/model/test/teardown";
+import ModelTestParentPrevent from "../../test/jest/helpers/model/test/parent_prevent";
 
 let baseWorld: BaseWorld | undefined;
