+++ test/jest/helpers/model/test/parent_prevent.ts
@@ -1,4 +1,4 @@
-import BaseWorld from "../../../jest/support/base_world";
+import BaseWorld from "../../../../jest/support/base_world";
 import ModelActions from "../actions";
 
 export default class ModelTestParentPrevent {
